refactored build process
